Reduced the memory footprint of some classes.

+++ source/graphics/2d/Textures/Char/CharSet.h
@@ -77,7 +77,7 @@ class CharSet : ListenerObject
 	// Charset spec
 	CharSetSpec* charSetSpec;
 	// Array spec of the charSet
-	uint32 tilesDisplacement;
+	uint16 tilesDisplacement;
 	// Memory displacement
 	uint16 offset;
 	// Written flag
